requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
face_recognition
python-dotenv
Pillow
numpy
Flask-CORS